Update examine.py
Updated examine, so examine() is accessible again to the users. However, I did not fix the issue with the ports yet. Only an error message was included for this. Also old dash references were removed and completely replaced with jupyter_dash. `mode` is now passed to run_server indicating if it will be shown inline or in a new jupyterlab tab.